+++ tests/MaxMind/Test/MinFraudTest.php
@@ -1102,7 +1102,9 @@ class MinFraudTest extends ServiceClientTester
         return [
             ['account_creation'],
             ['account_login'],
+            ['credit_application'],
             ['email_change'],
+            ['fund_transfer'],
             ['password_reset'],
             ['payout_change'],
             ['purchase'],