+++ CHANGELOG.md
@@ -11,6 +11,8 @@ CHANGELOG
   transaction.
 * Added the input `/payment/method`. This is the payment method associated
   with the transaction.
+* Updated `geoip2/geoip2` to 3.3.0, which includes new anonymizer and IP risk
+  outputs.
 
 3.3.0 (2025-05-23)
 ------------------